+++ dine/dine-core/src/main/java/com/dine/repository/SellerInfoRepository.java
@@ -4,8 +4,8 @@ import com.dine.repository.entity.SellerInfo;
 import com.dine.repository.support.LambdaJpaRepository;
 
 
-public interface SellerInfoRepository extends LambdaJpaRepository<SellerInfo, Integer> {
+public interface SellerInfoRepository extends LambdaJpaRepository<SellerInfo, Long> {
     SellerInfo findByPhone(String phone);
 
-    SellerInfo findBySellerId(Integer sellerId);
+    SellerInfo findBySellerId(Long sellerId);
 }
